// src/app/campaigns-management/models/upload.models.ts
export class Upload {

    name?: string;
    image_ref?: string;
    width?: number;
    height?: number;
    src?: any;
    image_url?: string;
    createdAt?: any;
    createdBy?: any;
    owner?: any;
}

export class UploadResponsiveDisplay {

    name?: string;
    image_ref?: string;
    width?: number;
    height?: number;
    src?: any;
    image_url?: string;
    image_json?: any;
    description?: string;
    brand?: string;
    createdAt?: any;
    createdBy?: any;
    owner?: any;
}